Integrate forging of OCSP status requests
ratched now has the capability to create its own (positive) OCSP
responses if the client queries for an OCSP status request. It will
create an OCSP ticket on-the-fly (if it has access to the CA key), sign
it and return it to the client. Also added debugging functionality for
memory blobs (such as DER-encoded OCSP tickets) and simplified
error-stack cleanup handling.