Massive refactoring of program options and server code
Massively refactored the way options are parsed. This consolidates the
way that arguments for default interception (i.e., with no specific SNI
target in mind) and specific interception (i.e., with --intercept
argument) are used. They now behave exactly the same way.

The errstack is, in principle, a good idea. However, due to the way it
is implemented (using enums), it creates much too heavy
interdependendies between modules.  This way of implementation was a
really stupid choice and needs to be refactored in the future (by
supplying a callback function instead of a enum value and have the local
modules spread out).

Currently not 100% sure if everything still works as expected. Consider
this work in progress.