Flakey integration tests finally fixed
A week of debugging. So much headache and confusion. Here's the
solution: OpenSSL s_client/s_server doesn't forward traffic unmodified,
but interprets the data (damn in-band signaling!). This means that when
the nonce started with a letter "q", it would "q"uit the connection.
Aaaaarghhhhhh! All this time I thought my timing was racy or off when
the solution was so, so, so, SO much simpler. Anyhow, fixed now by
prefixing all nonces with a letter "+".