gnuTLS integration tests now working
gnuTLS, at least in versions 3.5.8 and 3.6.1, requires for successful
OCSP ticket verification the the signer of that OCSP ticket is included
in the certificate chain sent by the TLS server. It is not sufficient
(at least for gnutls-cli) to have that CA certificate present in the
local CA trust store. For now, as a workaround, simply make ratched send
that root CA certificate in its chain.